+++ CHANGELOG.md
@@ -9,9 +9,14 @@ Packaged by Steven Czerwinski <czerwin@scalyr.com> on Jun 18, 2015 13:00 -0700
 
 Note, this release is still under development.  Release date and details are TBD.
 
+New features:
+
+* Allow any monitor's metric log's rate limiter to be modified via configuration, along with flush aggregation
+
 Bug fixes:
 
 * Do not fail Scalyr Agent service start if configuration file registry is not set (Windows)
+* Remove rate limiter on metric log for graphite monitor and add in flush aggregation
 
 ## 2.0.10 "Jumpy Jaguar" - June  9, 2015